Initial Markdown-ification of site content.
The content of the site is now written in Markdown rather than raw HTML.
At compile time, a custom Bazel rule runs a small tool (based on the
gomarkdown/markdown package) that renders the source markdown into HTML,
substituting it into a template. The generated HTML doesn't quite match
the hand-written HTML, but Chrome renders them pixel-for-pixel equally.

There is more work to be done, but this is a good stopping point.
Further work:
  * Markdownify the flower calculator template.
  * Allow the HTML title to be controlled somehow.