Markdown-ify flower breeding tool.
This was a lot easier than expected, since I can just embed HTML in
Markdown--and, it turns out, I can also embed Go template logic. I had
to disable MathJax since otherwise Markdown would interpret the
template's dollar sign as a mathematical expression. There are a couple
of roughly single-pixel differences in layout from the old
implementation; I'm fine with that. A slightly bigger issue is that the
customized title is now gone; I'll be thinking about how to fix that
next.