Add support for customizing page titles.
This is accomplished by adding a new flag to the Markdown renderer tool
to customize the title; the tool replaces a new "title marker" with the
specified title. The titles are plumbed through to the tool via new
attributes on the Bazel build rule which set the default title & any
override titles.